Pass complete state to scatterPlot #38
