+++ src/javascript/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import { chart } from "./functions";
+import { scatterPlot } from "./functions";
 import Suggestions from "./components/Suggestions.jsx";
 import Chart from "./components/Chart.jsx";
 
@@ -15,4 +15,4 @@ document.addEventListener("DOMContentLoaded", () => {
   }
 });
 
-window.chart = chart;
+window.scatterPlot = scatterPlot;
